Add hotfixes used after 4th November demo start (#24)
* Actor clearing fix

* Fix Whiterun exit crash

* Fix issue 11

+++ src/front/worldCleaner.ts
@@ -18,8 +18,15 @@ function processOneActor(): void {
   if (!actor || actorId === 0x14 || actor.isDisabled() || actor.isDeleted())
     return;
 
+  if (actor.isInDialogueWithPlayer()) {
+    // Deleting actor in dialogue crashes Skyrim
+    // https://github.com/skyrim-multiplayer/issue-tracker/issues/13
+    actor.setPosition(0, 0, 0);
+    return;
+  }
   actor.disable(false).then(() => {
     const ac = Actor.from(Game.getFormEx(actorId));
+    if (ac.getDialogueTarget() != null) return;
     if (ac) ac.delete();
   });
 }
